+++ packages/cli/src/config/find-config.js
@@ -6,14 +6,16 @@ const log = require('@home-gallery/logger')('cli.config.find')
 
 const isWindows = os.platform == 'win32'
 const env = process.env
+const cwd = path.resolve(process.cwd())
+const homeDir = os.homedir() || cwd
 const galleryDir = path.dirname(process.argv[1])
 
 const configFiles = ['gallery.config.yml', 'gallery.config.json']
 
 const defaultConfigFilename = 'gallery.config.yml'
-const defaultConfigDir = isWindows ? path.join(env.LOCALAPPDATA, 'home-gallery', 'config') : path.join(env.HOME, '.config', 'home-gallery')
+const defaultConfigDir = isWindows ? path.join(homeDir, 'home-gallery', 'config') : path.join(homeDir, '.config', 'home-gallery')
 const defaultConfigFile = path.join(defaultConfigDir, defaultConfigFilename)
-const defaultCacheDir = isWindows ? path.join(env.LOCALAPPDATA, 'home-gallery') : path.join(env.HOME, '.cache')
+const defaultCacheDir = isWindows ? path.join(homeDir, 'home-gallery') : path.join(homeDir, '.cache')
 
 const expandFiles = (dir, configPrefix = '') => configFiles.map(file => path.join(dir, `${configPrefix}${file}`))
 
@@ -21,10 +23,10 @@ const findConfig = async () => {
   const lookupFiles = [
     ...(env.GALLERY_CONFIG ? [env.GALLERY_CONFIG] : []),
     ...(env.GALLERY_CONFIG_DIR ? expandFiles(env.GALLERY_CONFIG_DIR) : []),
-    ...expandFiles(process.cwd()),
+    ...expandFiles(cwd),
     ...expandFiles(galleryDir),
     ...expandFiles(defaultConfigDir),
-    ...expandFiles(os.homedir(), '.'),
+    ...expandFiles(homeDir, '.'),
     ...(!isWindows ? expandFiles(path.resolve('/', 'etc')) : [])
   ]
   log.trace(`Lookup configuration in ${lookupFiles.join(', ')}`)