+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- cli: Fix find config for undefined HOME env
 - extractor: Fix preview video size of portrait videos
 - database: Fix video rotation
 - database: Fix video duration